refactor(cli): extract the hosted control-plane client into a shared cloudapi package
